Moved the tofi folder to sdk/common/adi.
